Task4/app/requirements.txt
flask
psycopg2-binary
selenium